refactor(app): collapse App.tsx to thin shell, extract orthogonal hooks (#38)
App.tsx drops from 2,643 → 803 lines (-70%). Every useEffect, handler,
and ref now belongs to a hook. App.tsx wires the 9 prescribed hooks in
spec order (useZoomAndTheme → useShellConsent → useProjects → useTabs →
useTabNavigation → useExtensionsHydration → useBridgeMessages →
useKeyboardShortcuts → useWindowApi) interleaved with 9 orthogonal hooks
extracted from residual App-level logic.

New orthogonal hooks:
  - useBootConfig — config.toml load + theme/zoom/sidebar-width restore
  - useChat — sendChat, setModel, stopPrompt, append helpers, export markdown
  - useFocus — terminal/sidebar toggles + composer/terminal focus shuttle
  - useFrontendStateMirror — debounced bridge state mirror
  - useNotifications — toast stack + OS completion notification
  - useOsEdges — shell-output/exit/title, supervisor signals, menu, dnd, paste
  - useProjectOps — project list ops + per-project tab buckets
  - useUiOverlays — palette + settings + search modal trio
  - useUpdater — Tauri updater check-and-install

Other moves:
  - routeShellWrite into useShellConsent (security-domain pairing)
  - recomputeModelPicker → src/utils/modelPicker.ts (pure utility, dropped
    from useTabs / useProjectOps / BridgeMessageContext plumbing)

App.tsx forward-handle refs (projectOpsHandleRef, projectsHandleRef,
pushNotificationRef, chatActionsRef, stateRef-mirror) wired in commit
phase via useEffect to satisfy react-hooks/refs (no render-time ref
mutation).

Check gate: tsc clean, eslint 0 warnings, 583/583 vitest, 0 clippy
warnings, 74/74 cargo tests. aethon-debug e2e: 3 tabs (agent + agent +
shell), project-bucket round-trip preserves tabs, agent hot-reload via
touch agent/main.ts cleanly cycles the supervisor while the frontend
stays connected.
